refactor: rename `objects/cmd` sub packages
