+++ services/objects/cmd/messaging/server.go
@@ -1,4 +1,4 @@
 // this package would be responsible for polling the messages table
 // and publishing the related events, commands, ... . It will also listen
 // to incoming messages and call the appropriate handler.
-package kafka
+package messaging
